folder organization. removed unecessary components
