--- components/LabDataDrawer.tsx
@@ -1,2 +1,0 @@
-
-//drawer component. displays cards with test data when test chart is clicked
