+++ components/charts/radialcharts/DailyProtein.tsx
@@ -16,8 +16,8 @@ import {
   CardFooter,
   CardHeader,
   CardTitle,
-} from "@/components/ui/card"
-import { ChartConfig, ChartContainer } from "@/components/ui/chart"
+} from "@/components/ui_primitives/card"
+import { ChartConfig, ChartContainer } from "@/components/ui_primitives/chart"
 const chartData = [
   { metric: "todayprotein", number: 85, fill: "var(--color-metric)" },
 ]
