+++ components/charts/radialcharts/UniversalRingChart.tsx
@@ -1,7 +1,7 @@
 "use client"
 import { Label, PolarGrid, PolarRadiusAxis, RadialBar, RadialBarChart } from "recharts"
-import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui/card"
-import { ChartConfig, ChartContainer } from "@/components/ui/chart"
+import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui_primitives/card"
+import { ChartConfig, ChartContainer } from "@/components/ui_primitives/chart"
 import { useEffect, useState } from "react"
 import { GenericRingChart } from "@/utils/types"
 import { useIsMobile } from "@/hooks/use-mobile"
