+++ components/charts/radialcharts/WeightChart.tsx
@@ -1,6 +1,5 @@
 "use client"
 
-import { TrendingUp } from "lucide-react"
 import {
   Label,
   PolarGrid,
@@ -16,10 +15,11 @@ import {
   CardFooter,
   CardHeader,
   CardTitle,
-} from "@/components/ui/card"
-import { ChartConfig, ChartContainer } from "@/components/ui/chart"
+} from "@/components/ui_primitives/card"
+import { ChartConfig, ChartContainer } from "@/components/ui_primitives/chart"
+import { useIsMobile } from "@/hooks/use-mobile"
 const chartData = [
-  { metric: "todaycalories", number: 1850, fill: "var(--color-metric)" },
+  { metric: "weightlost", number: 20, fill: "var(--color-metric)" },
 ]
 
 const chartConfig = {
@@ -28,35 +28,37 @@ const chartConfig = {
   },
   metric: {
     label: "Weight Lost",
-    color: "hsl(var(--chart-2))",
+    color: "hsl(var(--chart-4))",
   },
 } satisfies ChartConfig
 
-export function DailyCalorieRingChart() {
+export function WeightChart() {
+  const isMobile = useIsMobile();
+  let percentage = 50;
   return (
-    <Card className="flex flex-col max-w-fit mb-4 rounded-xl bg-gradient-to-br from-teal-900/10 to-neutral-950">
+    <Card className="flex flex-col align-center w-full mb-4 rounded-xl bg-gradient-to-br from-blue00/10 to-neutral-950">
       <CardHeader className="items-center pb-0">
-        <CardTitle>{`Today's Calorie Intake`}</CardTitle>
-        <CardDescription>Intake compared to maintenence</CardDescription>
+        <CardTitle className="mb-4 ">Weight Loss Progress</CardTitle>
+        <CardDescription className="hidden md:relative">Total lost out of goal</CardDescription>
       </CardHeader>
       <CardContent className="flex-1 pb-0">
         <ChartContainer
           config={chartConfig}
-          className="mx-auto aspect-square max-h-[250px]"
+          className="mx-0 md:mx-auto aspect-square max-h-[220px] mb-4"
         >
           <RadialBarChart
             data={chartData}
-            startAngle={270}
-            endAngle={0}
-            innerRadius={80}
-            outerRadius={110}
+            startAngle={Math.abs(percentage * 3.6 - 270)}
+            endAngle={270}
+            innerRadius={isMobile ? 54 : 80}
+            outerRadius={isMobile ? 72 : 110}
           >
             <PolarGrid
               gridType="circle"
               radialLines={false}
               stroke="none"
               className="first:fill-muted last:fill-background"
-              polarRadius={[86, 74]}
+              polarRadius={isMobile ? [58, 50] : [86, 74]}
             />
             <RadialBar dataKey="number" background cornerRadius={10} />
             <PolarRadiusAxis tick={false} tickLine={false} axisLine={false}>
@@ -82,7 +84,7 @@ export function DailyCalorieRingChart() {
                           y={(viewBox.cy || 0) + 24}
                           className="fill-muted-foreground"
                         >
-                          Calories
+                          Pounds
                         </tspan>
                       </text>
                     )
@@ -93,9 +95,9 @@ export function DailyCalorieRingChart() {
           </RadialBarChart>
         </ChartContainer>
       </CardContent>
-      <CardFooter className="flex-col gap-2 text-sm">
+      <CardFooter className="flex-col gap-2 pt-4 text-sm">
         <div className="flex items-center gap-2 font-light text-neutral-500 leading-none">
-          Your maintenence calorie intake is 2200
+          Only 25 pounds left to reach your goal
         </div>
       </CardFooter>
     </Card>
