+++ components/global/Footer.tsx
