+++ components/gym/WorkOutForm.tsx
@@ -4,14 +4,14 @@ import React, { useState, useEffect } from "react"
 import { useForm, useFieldArray, FormProvider } from "react-hook-form"
 import { zodResolver } from "@hookform/resolvers/zod"
 import { z } from "zod"
-import { Button } from "@/components/ui/button"
-import { Form, FormControl, FormField, FormItem, FormLabel } from "@/components/ui/form"
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
-import { Checkbox } from "@/components/ui/checkbox"
+import { Button } from "@/components/ui_primitives/button"
+import { Form, FormControl, FormField, FormItem, FormLabel } from "@/components/ui_primitives/form"
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui_primitives/select"
+import { Checkbox } from "@/components/ui_primitives/checkbox"
 import { addItemToTable } from "@/actions/AWS"
 import StrengthExercise from "@/components/gym/strength-exercise"
 import CardioExercise from "@/components/gym/CardioExerciseEntry"
-import { Accordion, AccordionContent, AccordionTrigger, AccordionItem } from "../ui/accordion"
+import { Accordion, AccordionContent, AccordionTrigger, AccordionItem } from "../ui_primitives/accordion"
 
 export const dynamic = 'force-dynamic';
