+++ components/ui_primitives/alert-dialog.tsx
@@ -4,7 +4,7 @@ import * as React from "react"
 import * as AlertDialogPrimitive from "@radix-ui/react-alert-dialog"
 
 import { cn } from "@/lib/utils"
-import { buttonVariants } from "@/components/ui/button"
+import { buttonVariants } from "@/components/ui_primitives/button"
 
 const AlertDialog = AlertDialogPrimitive.Root
