+++ components/ui_primitives/toaster.tsx
@@ -8,7 +8,7 @@ import {
   ToastProvider,
   ToastTitle,
   ToastViewport,
-} from "@/components/ui/toast"
+} from "@/components/ui_primitives/toast"
 
 export function Toaster() {
   const { toasts } = useToast()
